# client/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eyened API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/form-annotations": {
      "post": {
        "summary": "Create Form Annotation",
        "description": "Create a form annotation.",
        "operationId": "create_form_annotation_form_annotations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormAnnotationPUT"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAnnotationGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Form Annotations",
        "description": "List active form annotations, optionally filtered.",
        "operationId": "get_form_annotations_form_annotations_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "name": "study_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Study Id"
            }
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Image Id"
            }
          },
          {
            "name": "form_schema_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Form Schema Id"
            }
          },
          {
            "name": "sub_task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sub Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormAnnotationGET"
                  },
                  "title": "Response Get Form Annotations Form Annotations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/form-annotations/{annotation_id}": {
      "get": {
        "summary": "Get Form Annotation",
        "description": "Get a single form annotation by id.",
        "operationId": "get_form_annotation_form_annotations__annotation_id__get",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAnnotationGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Form Annotation",
        "description": "Partially update a form annotation.",
        "operationId": "update_form_annotation_form_annotations__annotation_id__patch",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormAnnotationPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAnnotationGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Form Annotation",
        "description": "Soft-delete a form annotation.",
        "operationId": "delete_form_annotation_form_annotations__annotation_id__delete",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/form-annotations/{form_annotation_id}/value": {
      "get": {
        "summary": "Get Form Annotation Value",
        "description": "Get a form annotation's raw FormData payload.",
        "operationId": "get_form_annotation_value_form_annotations__form_annotation_id__value_get",
        "parameters": [
          {
            "name": "form_annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Form Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Form Annotation Value",
        "description": "Overwrite a form annotation's FormData payload.",
        "operationId": "update_form_annotation_value_form_annotations__form_annotation_id__value_put",
        "parameters": [
          {
            "name": "form_annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Form Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/form-annotations/{annotation_id}/tags": {
      "post": {
        "summary": "Tag Form Annotation",
        "description": "Attach a Tag to a FormAnnotation by tag ID (idempotent).",
        "operationId": "tag_form_annotation_form_annotations__annotation_id__tags_post",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPOST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/form-annotations/{annotation_id}/tags/{tag_id}": {
      "delete": {
        "summary": "Untag Form Annotation",
        "description": "Remove a Tag from a FormAnnotation (idempotent).",
        "operationId": "untag_form_annotation_form_annotations__annotation_id__tags__tag_id__delete",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Form Annotation Tag",
        "description": "Update comment on an existing FormAnnotation tag link.",
        "operationId": "patch_form_annotation_tag_form_annotations__annotation_id__tags__tag_id__patch",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Annotation Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatorMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CreatorMeta"
      },
...
      "FormAnnotationGET": {
        "properties": {
          "form_schema_id": {
            "type": "integer",
            "title": "Form Schema Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "study_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Id"
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "laterality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Laterality"
              },
              {
                "type": "null"
              }
            ]
          },
          "sub_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Task Id"
          },
          "form_data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Form Data"
          },
          "form_annotation_reference_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Annotation Reference Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "annotation_type": {
            "type": "string",
            "const": "grader_form",
            "title": "Annotation Type"
          },
          "object_type": {
            "type": "string",
            "enum": [
              "patient",
              "study",
              "image_instance"
            ],
            "title": "Object Type"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagMeta"
            },
            "type": "array",
            "title": "Tags"
          },
          "creator": {
            "$ref": "#/components/schemas/CreatorMeta"
          },
          "date_inserted": {
            "type": "string",
            "format": "date-time",
            "title": "Date Inserted"
          },
          "date_modified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Modified"
          }
        },
        "type": "object",
        "required": [
          "form_schema_id",
          "patient_id",
          "id",
          "annotation_type",
          "object_type",
          "tags",
          "creator",
          "date_inserted"
        ],
        "title": "FormAnnotationGET"
      },
      "FormAnnotationPATCH": {
        "properties": {
          "form_schema_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Schema Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "study_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Id"
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "laterality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Laterality"
              },
              {
                "type": "null"
              }
            ]
          },
          "sub_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Task Id"
          },
          "form_data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Form Data"
          },
          "form_annotation_reference_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Annotation Reference Id"
          }
        },
        "type": "object",
        "title": "FormAnnotationPATCH",
        "description": "Partial update for FormAnnotation with optional fields."
      },
      "FormAnnotationPUT": {
        "properties": {
          "form_schema_id": {
            "type": "integer",
            "title": "Form Schema Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "study_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Id"
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "laterality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Laterality"
              },
              {
                "type": "null"
              }
            ]
          },
          "sub_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Task Id"
          },
          "form_data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Form Data"
          },
          "form_annotation_reference_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Annotation Reference Id"
          }
        },
        "type": "object",
        "required": [
          "form_schema_id",
          "patient_id"
        ],
        "title": "FormAnnotationPUT"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Laterality": {
        "type": "string",
        "enum": [
          "L",
          "R"
        ],
        "title": "Laterality"
      },
...
      "ObjectTagPATCH": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "ObjectTagPATCH"
      },
      "ObjectTagPOST": {
        "properties": {
          "tag_id": {
            "type": "integer",
            "title": "Tag Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "tag_id"
        ],
        "title": "ObjectTagPOST"
      },
...
      "TagMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tagger": {
            "$ref": "#/components/schemas/CreatorMeta"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "tagger",
          "date"
        ],
        "title": "TagMeta"
      },
...
    }
  }
}